main.py
import discord
from discord.ext import commands

bot = commands.Bot(command_prefix='!')

@bot.event
async def on_ready():
    print(f'Logged in as {bot.user.name}')

@bot.command()
async def schedule(ctx, from_date, to_date):
    # from_dateとto_dateの間の水曜日、土曜日、日曜日の日付を取得
    # ここで、from_dateとto_dateの指定された日付を処理して、目的の日付のリストを作成する必要があります

    wednesdays = []  # 水曜日の日付リスト
    saturdays = []  # 土曜日の日付リスト
    sundays = []    # 日曜日の日付リスト

    # 出欠確認メッセージの作成
    message = "出欠確認\n"
    for date in wednesdays:
        message += f"{date.strftime('%Y-%m-%d')} (水曜日): ✅\n"
    for date in saturdays:
        message += f"{date.strftime('%Y-%m-%d')} (土曜日): ✅\n"
    for date in sundays:
        message += f"{date.strftime('%Y-%m-%d')} (日曜日): ✅\n"

    # スタンプを送信して出欠確認
    await ctx.send(message)
    await ctx.message.add_reaction("✅")

# Discordボットを起動
bot.run('YOUR_BOT_TOKEN')

